Bug fixes for creating posts and logging out.
Fixes a bug where the posts without a title would cause the application
to fail due to the logic in determining if a title was present.

Also fixes a bug where it would be possible to attempt to log out with
None values for either the instance or the access_token. It should never
have been reached due to error checking in loading the config, but it
now passes the Pylint test.