Adds initial collection prompt and post confirmation.
Adds a prompt when first launching the TUI to prompt for the collection
since nothing else can be done until that has been specified. This would
result in an error message on any command other than login/logout which
would force the user to enter their selection twice.

Adds a check after the `EDITOR` has been closed to see if the user wants
to publish the recently authored post, edit it once more, or discard the
content.